# backend/Tempswagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Trilho API",
    "version": "v1"
  },
  "paths": {
...
    "/api/stations/{id}/crowd": {
      "get": {
        "tags": [
          "StationEndpoints"
        ],
        "operationId": "GetStationCrowd",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/stations/{id}/forecast": {
      "get": {
        "tags": [
          "StationEndpoints"
        ],
        "operationId": "GetStationForecast",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/auth/firebase": {
      "post": {
        "tags": [
          "UserEndpoints"
        ],
        "operationId": "FirebaseAuth",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FirebaseAuthRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FirebaseAuthRequestDto": {
        "type": "object",
        "properties": {
          "idToken": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}